fixed numbeer input
now allows negative and decimal numbers

btw manual.gamemaker.io is not completely correct

~~~
real

This function can be used to turn a value into a real number.

When using this function on a string, numbers, minus signs, decimal points and exponential parts in the string are taken into account, while other characters (such as letters) will cause an error to be thrown.
~~~
letters will not cause an error unless there aren't any numbers in the string at all.
(which wouldn't have been a problem if it did cause an error, as there is a way to handle errors: "The try, catch and finally statements can be used in your game for error checking and permit you to test out blocks of code and control what happens if any runtime exceptions occur. Using these will prevent the exception ending the game and showing the standard error message to the user, but this means that you will have to handle what happens next in that case")

so it works perfectly and takes only the first negative sign and the first decimal and the appropriate numbers, and ignores everything else. (instead of giving an error as the manual.gamemaker.io says)

so this piece of code is perfect:
if string_digits(keyboard_string)!=""